+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,0 +1,32 @@
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardBundle;
+import blackjack.domain.game.Score;
+
+public abstract class Participant {
+
+    protected final String name;
+    protected CardBundle cardBundle;
+
+    protected Participant(final String name, final CardBundle cardBundle) {
+        this.name = name;
+        this.cardBundle = cardBundle;
+    }
+
+    public abstract void receiveCard(Card card);
+
+    public abstract boolean canReceive();
+
+    public Score getCurrentScore() {
+        return cardBundle.getScore();
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public CardBundle getCardBundle() {
+        return cardBundle;
+    }
+}
